Using a default class called 'form' on form helpers
The legacy css was using generic selectors for form elements, e.g.
input[type="text"] which was adding unwanted styles to new form layouts.
This commit monkey patches the form helpers to set a default css class
of 'form' on form helper elements, and the form scss is now scoped with
that class, so existing forms ought to work as before and new forms
start with clean css. Once all the forms have been migrated to the style
guide then these files can be deleted.